requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
crispy-bootstrap5==0.6
cryptography==37.0.2
Django==3.2.13
django-crispy-forms==1.14.0
django-environ==0.8.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-pwa==1.0.10
django-webpush==0.3.4
http-ece==1.1.0
idna==3.3
jsonfield==3.1.0
py-vapid==1.8.2
pycparser==2.21
pytz==2022.1
pywebpush==1.9.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.9